+++ src/main/java/com/sky/pro/HW5/exception/WrongLoginException.java
@@ -1,14 +1,6 @@
-
-package exception;
-
-
+package com.sky.pro.HW5.exception;
 public class WrongLoginException extends RuntimeException {
-
     public WrongLoginException(String message) {
-
         super(message);
-
-
-
     }
 }
