+++ src/main/java/com/sky/pro/HW5/exception/WrongPasswordException.java
@@ -1,14 +1,6 @@
-
-package exception;
-
-
+package com.sky.pro.HW5.exception;
 public class WrongPasswordException extends RuntimeException {
-
     public WrongPasswordException(String message) {
-
-
         super(message);
-
-
     }
 }
