+++ src/test/java/com/sky/pro/HW5/Hw5ApplicationTests.java
@@ -1,0 +1,13 @@
+package com.sky.pro.HW5;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Hw5ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
